refactor: bag operation
